requirements.txt
streamlit
os
google.generativeai